FIXATA CLASSIFICA

+++ php/game/quit_game.php
@@ -10,12 +10,23 @@
 	$match = $_POST["id"];
 	$user  = $_SESSION["session_mail"];
 
-	// eliminiamo il giocatore dalla tabella giocatori
-	$query= "delete from `player` where `match_id` = :match and user = :user";
-	$request = $pdo->prepare($query);
-	$request->bindParam(':match', $match, PDO::PARAM_STR);
-	$request->bindParam(':user',   $user, PDO::PARAM_STR);
-	$a = $request->execute();
 
-	echo $a;
+	$query= "select status from `match` where id = :id";
+	$request = $pdo->prepare($query);
+	$request->bindParam(':id', $match, PDO::PARAM_STR);
+	$request->execute();
+	$status = $request->fetch()["status"];
+
+	if ($status == 0){
+		// eliminiamo il giocatore dalla tabella giocatori
+		$query= "delete from `player` where `match_id` = :match and user = :user";
+		$request = $pdo->prepare($query);
+		$request->bindParam(':match', $match, PDO::PARAM_STR);
+		$request->bindParam(':user',   $user, PDO::PARAM_STR);
+		$a = $request->execute();
+
+		echo $a;
+	}
+
+	
 ?>
